app/health: add remaining check tests
